refactor: move K8sStatusError and isK8sStatus to @odh-dashboard/k8s-core
Centralizes K8sStatusError class and isK8sStatus type guard in
k8s-core as the canonical shared location. Updates all consumers
across model-serving, nim-serving, and the frontend host to import
directly from k8s-core instead of @odh-dashboard/internal.

Part of RHOAIENG-65103 (child task 1).
