+++ frontend/src/api/k8s/__tests__/groups.spec.ts
@@ -1,9 +1,9 @@
 import { testHook } from '@odh-dashboard/jest-config/hooks';
+import { K8sStatusError } from '@odh-dashboard/k8s-core';
 import { groupVersionKind, useAccessReview, useGroups } from '#~/api';
 import { GroupModel } from '#~/api/models';
 import { mockGroup } from '#~/__mocks__/mockGroup';
 import { mock404Error, mock500Error } from '#~/__mocks__/mockK8sStatus';
-import { K8sStatusError } from '#~/api/errorUtils';
 import useK8sWatchResourceList from '#~/utilities/useK8sWatchResourceList';
 import { GroupKind } from '#~/k8sTypes';
