+++ frontend/src/pages/projects/screens/detail/notebooks/NotebookUpdateImageModalFooter.tsx
@@ -1,6 +1,7 @@
 import { ModalFooter, Button, StackItem, Alert, Stack } from '@patternfly/react-core';
 import * as React from 'react';
-import { K8sStatusError, patchNotebookImage } from '#~/api';
+import { K8sStatusError } from '@odh-dashboard/k8s-core';
+import { patchNotebookImage } from '#~/api';
 import { fireFormTrackingEvent } from '#~/concepts/analyticsTracking/segmentIOUtils';
 import {
   FormTrackingEventProperties,
