+++ packages/k8s-core/src/index.ts
@@ -139,6 +139,7 @@ export type {
 } from './connectionTypes';
 
 export { NamespaceApplicationCase, byName, namespaceToProjectDisplayName } from './projectUtils';
+export { isK8sStatus, K8sStatusError } from './errorUtils';
 
 export {
   isSecretKind,
